refactor: fix typo in betaUtil
